renamed randomfact to homepage and moved into components
