+++ src/components/Home.jsx
@@ -1,13 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import './App.css';
+import '../App.css';
 import { Link } from 'react-router-dom';
-import QLogo from './assets/quantumicon.png';
+import QLogo from '../assets/quantumicon.png';
 import Button from '@mui/material/Button';
 import CircularProgress from '@mui/material/CircularProgress';
-import MapComponent from './MapComponent.jsx';
-import AddressSearch from './components/AddressSearch.jsx';
-import UploadRunsheet from './components/UploadRunsheet.jsx';
-import * as Constants from './Constants.js';
+import MapComponent from '../MapComponent.jsx';
+import AddressSearch from './AddressSearch.jsx';
+import * as Constants from '../Constants.js';
 
 const FactGenerator = () => {
     const [fact, setFact] = useState('');
